Refactor TaskForm
