+++ src/components/forms/TaskCreateForm.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 
 // mui
-import { Box, Button, Grid } from '@mui/material';
+import { Box, Button } from '@mui/material';
 import AddIcon from '@mui/icons-material/Add';
 
 // store
@@ -9,12 +9,8 @@ import { useStore } from 'store/useStore';
 import { ADD_TASK } from 'store/actions';
 
 // components
-import PriorityControl from './PriorityControl';
-import DueDateControl from './DueDateControl';
-import TextControl from './TextControl';
 import WarningDialog from 'components/WarningDialog';
-import CustomForm from './CustomForm';
-import LabelControl from './LabelControl';
+import TaskForm from './TaskForm';
 
 // helpers
 import isEmpty from 'lodash/isEmpty';
@@ -26,70 +22,6 @@ const AddButton = ({ openForm }) => (
     Add task
   </Button>
 );
-
-// renders the task creation form's ui
-const Form = ({ createItem, closeForm, updateData, formData, onSubmit }) => {
-  // extract values from state
-  const { name, description, priority, due, label } = formData;
-
-  // create setters
-  const setName = updateData('name');
-  const setDescription = updateData('description');
-  const setPriority = updateData('priority');
-  const setDue = updateData('due');
-  const setLabel = updateData('label');
-
-  const handleSubmit = () => {
-    createItem();
-    onSubmit();
-  };
-
-  const update = (setter) => (e) => setter(e.target.value);
-
-  return (
-    <CustomForm
-      onSubmit={handleSubmit}
-      title="Add a Task"
-      canSubmit={name}
-      onCancel={closeForm}
-      submitButton="Add Task"
-    >
-      <TextControl
-        name="Task"
-        value={name || ''}
-        onChange={update(setName)}
-        required
-        autoFocus
-      />
-      <TextControl
-        name="Description"
-        value={description || ''}
-        onChange={update(setDescription)}
-        lines={3}
-      />
-      <Grid
-        container
-        justifyContent="space-between"
-        alignItems="center"
-        spacing={2}
-      >
-        <Grid item>
-          <DueDateControl date={due} setDate={setDue} />
-        </Grid>
-        <Grid item>
-          <Grid container alignItems="center" spacing={2}>
-            <Grid item>
-              <LabelControl label={label} setLabel={setLabel} />
-            </Grid>
-            <Grid item>
-              <PriorityControl priority={priority} setPriority={setPriority} />
-            </Grid>
-          </Grid>
-        </Grid>
-      </Grid>
-    </CustomForm>
-  );
-};
 
 // main component control: maintains form state, handles dispatch to store
 const TaskCreateForm = ({ defaultItem }) => {
@@ -110,17 +42,6 @@ const TaskCreateForm = ({ defaultItem }) => {
 
   const newItem = () => dispatch(itemCreator(formData));
 
-  // accepts a string representing a task property,
-  // returns a function that accepts a value, which updates the state as: state[property] = value;
-  const updateData = (property) => (value) => {
-    setFormData((prev) => {
-      return {
-        ...prev,
-        [property]: value,
-      };
-    });
-  };
-
   // clears the entire form
   const clearForm = () => setFormData({ ...defaultItem });
 
@@ -139,15 +60,19 @@ const TaskCreateForm = ({ defaultItem }) => {
     else setWarningOpen(true);
   };
 
+  const handleSubmit = () => {
+    newItem();
+    confirmCloseForm();
+  };
+
   return (
     <Box sx={{ mt: 2 }}>
       {formOpen ? (
-        <Form
-          createItem={newItem}
+        <TaskForm
+          data={formData}
+          setter={setFormData}
           closeForm={closeForm}
-          updateData={updateData}
-          formData={formData}
-          onSubmit={confirmCloseForm}
+          onSubmit={handleSubmit}
         />
       ) : (
         <AddButton openForm={() => setFormOpen(true)} />
